+++ src/test/java/net/pwall/util/TestStrings.java
@@ -98,4 +98,54 @@ public class TestStrings {
                 Strings.toEnglish(0x80000000));
     }
 
+    @Test
+    public void test_appendInt() throws IOException {
+        StringBuilder sb = new StringBuilder();
+        int i = 123456789;
+        Strings.appendInt(sb, i);
+        assertEquals(String.valueOf(i), sb.toString());
+        sb = new StringBuilder();
+        i = -1000;
+        Strings.appendInt(sb, i);
+        assertEquals(String.valueOf(i), sb.toString());
+        sb = new StringBuilder();
+        i = Integer.MAX_VALUE;
+        Strings.appendInt(sb, i);
+        assertEquals(String.valueOf(i), sb.toString());
+        sb = new StringBuilder();
+        i = Integer.MIN_VALUE;
+        Strings.appendInt(sb, i);
+        assertEquals(String.valueOf(i), sb.toString());
+        for (i = -20000; i < 20000; i++) {
+            sb = new StringBuilder();
+            Strings.appendInt(sb, i);
+            assertEquals(String.valueOf(i), sb.toString());
+        }
+    }
+
+    @Test
+    public void test_appendLong() throws IOException {
+        StringBuilder sb = new StringBuilder();
+        long i = 123456789123456789L;
+        Strings.appendLong(sb, i);
+        assertEquals(String.valueOf(i), sb.toString());
+        sb = new StringBuilder();
+        i = -1000;
+        Strings.appendLong(sb, i);
+        assertEquals(String.valueOf(i), sb.toString());
+        sb = new StringBuilder();
+        i = Long.MAX_VALUE;
+        Strings.appendLong(sb, i);
+        assertEquals(String.valueOf(i), sb.toString());
+        sb = new StringBuilder();
+        i = Long.MIN_VALUE;
+        Strings.appendLong(sb, i);
+        assertEquals(String.valueOf(i), sb.toString());
+        for (i = -20000; i < 20000; i++) {
+            sb = new StringBuilder();
+            Strings.appendLong(sb, i);
+            assertEquals(String.valueOf(i), sb.toString());
+        }
+    }
+
 }